refactor: move drawer to toplevel
