+++ packages/renderer-react/src/plugins/react/drawer.tsx
@@ -1,0 +1,44 @@
+import { BlockType, FocusMode, OpType } from '@blink-mind/core';
+import { cancelEvent, Icon } from '@blink-mind/renderer-react';
+import { Drawer } from '@blueprintjs/core';
+import * as React from 'react';
+import styled from 'styled-components';
+const DescEditorWrapper = styled.div`
+  overflow: auto;
+  padding: 0px 0px 0px 20px;
+  background: #88888850;
+`;
+
+export function renderDrawer(props) {
+  const { controller, model, topicKey } = props;
+  if (model.focusMode === FocusMode.EDITING_DESC) {
+    const onDescEditorClose = e => {
+      e.stopPropagation();
+      const key = `topic-desc-data-${topicKey}`;
+      const descData = controller.run('deleteTempValue', { key });
+      controller.run('operation', {
+        ...props,
+        opType: OpType.SET_TOPIC_BLOCK,
+        topicKey,
+        blockType: BlockType.DESC,
+        data: descData,
+        focusMode: FocusMode.NORMAL
+      });
+    };
+    const descEditor = controller.run('renderTopicDescEditor', props);
+    return (
+      <Drawer
+        title="Edit Notes"
+        icon={Icon('note')}
+        isOpen
+        hasBackdrop
+        backdropProps={{ onMouseDown: cancelEvent }}
+        isCloseButtonShown={false}
+        onClose={onDescEditorClose}
+        size="70%"
+      >
+        <DescEditorWrapper>{descEditor}</DescEditorWrapper>
+      </Drawer>
+    );
+  }
+}
